@@ -6,15 +6,11 @@ use App\Events\PermisoEmpleadoEvent;
 use App\Events\PermisoNotificacionEvent;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\RecursosHumanos\NominaPrestamos\PermisoEmpleadoRequest;
-use App\Http\Resources\RecursosHumanos\NominaPrestamos\ArchivoPermisoEmpleadoResource;
 use App\Http\Resources\RecursosHumanos\NominaPrestamos\PermisoEmpleadoResource;
 use App\Models\Autorizacion;
 use App\Models\Empleado;
-use App\Models\Notificacion;
 use App\Models\RecursosHumanos\NominaPrestamos\PermisoEmpleado;
-use App\Models\User;
 use Carbon\Carbon;
-use eloquentFilter\QueryFilter\Queries\WhereIn;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -23,20 +19,24 @@ use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
 use Src\App\ArchivoService;
 use Src\Config\RutasStorage;
-use Src\Shared\GuardarArchivo;
 use Src\Shared\Utils;
 
 class PermisoEmpleadoController extends Controller
 {
-    private $entidad = 'PERMISO_EMPLEADO';
-    private $archivoService;
+    private string $entidad = 'PERMISO_EMPLEADO';
+    private ArchivoService $archivoService;
+
     public function __construct()
     {
         $this->archivoService = new ArchivoService();
         $this->middleware('can:puede.ver.permiso_nomina')->only('index', 'show');
         $this->middleware('can:puede.crear.permiso_nomina')->only('store');
     }
-    public function archivo_permiso_empleado(Request $request)
+
+    /**
+     * @throws Exception
+     */
+    public function archivoPermisoEmpleado(Request $request)
     {
         $request->validate([
             'permiso_id' => 'required|numeric|integer',
@@ -59,22 +59,22 @@ class PermisoEmpleadoController extends Controller
         $modelo = $this->archivoService->guardarArchivo($permiso_empleado, $request->file, RutasStorage::DOCUMENTOS_PERMISO_EMPLEADO->value . auth()->user()->empleado->identificacion);
 
 
-        // return response()->json(['modelo' => $permiso_empleado, 'mensaje' => 'Subido exitosamente!']);
         return response()->json(['modelo' => $modelo, 'mensaje' => 'Subido exitosamente!']);
     }
-    public function index_archivo_permiso_empleado(Request $request)
+
+    /**
+     * @throws Exception
+     */
+    public function indexArchivoPermisoEmpleado(Request $request)
     {
         $permiso = PermisoEmpleado::where('id', $request->permiso_id)->first();
-        // $results = ArchivoPermisoEmpleadoResource::collection($results);
         $results = $this->archivoService->listarArchivos($permiso);
         return response()->json(compact('results'));
     }
-    public function index(Request $request)
+
+    public function index()
     {
-        $results = [];
-        $usuario = Auth::user();
-        $usuario_ac = User::where('id', $usuario->id)->first();
-        if ($usuario_ac->hasRole('RECURSOS HUMANOS')) {
+        if (auth()->user()->hasRole('RECURSOS HUMANOS')) {
             $results = PermisoEmpleado::ignoreRequest(['campos'])->filter()->get();
         } else {
             $empleados = Empleado::where('jefe_id', Auth::user()->empleado->id)->orWhere('id', Auth::user()->empleado->id)->get('id');
@@ -84,74 +84,75 @@ class PermisoEmpleadoController extends Controller
         return response()->json(compact('results'));
     }
 
-    public function create(Request $request)
-    {
-        $permisoEmpleado = new PermisoEmpleado();
-        $permisoEmpleado->nombre = $request->nombre;
-        $permisoEmpleado->save();
-        return $permisoEmpleado;
-    }
+//    public function create(Request $request)
+//    {
+//        $permisoEmpleado = new PermisoEmpleado();
+//        $permisoEmpleado->nombre = $request->nombre;
+//        $permisoEmpleado->save();
+//        return $permisoEmpleado;
+//    }
 
     public function store(PermisoEmpleadoRequest $request)
     {
         try {
             $datos = $request->validated();
-            // if ($request->tieneDocumento == false) {
-            //     throw ValidationException::withMessages([
-            //         '404' => ['Debe seleccionar al menos un archivo.'],
-            //     ]);
-            // }
             DB::beginTransaction();
-            $datos['tipo_permiso_id'] =  $request->safe()->only(['tipo_permiso'])['tipo_permiso'];
-            $datos['estado_permiso_id'] =  PermisoEmpleado::PENDIENTE;
-            $permisoEmpleado = PermisoEmpleado::create($datos);
-            event(new PermisoEmpleadoEvent($permisoEmpleado));
-            $modelo = new PermisoEmpleadoResource($permisoEmpleado);
+            $datos['tipo_permiso_id'] = $request->safe()->only(['tipo_permiso'])['tipo_permiso'];
+            $datos['estado_permiso_id'] = PermisoEmpleado::PENDIENTE;
+            $permiso = PermisoEmpleado::create($datos);
+            event(new PermisoEmpleadoEvent($permiso));
+            $modelo = new PermisoEmpleadoResource($permiso);
             DB::commit();
             $mensaje = Utils::obtenerMensaje($this->entidad, 'store');
             return response()->json(compact('mensaje', 'modelo'));
         } catch (Exception $e) {
             DB::rollBack();
             Log::channel('testing')->info('Log', ['ERROR en el insert de permiso de empleado', $e->getMessage(), $e->getLine()]);
-            return response()->json(['mensaje' => 'Ha ocurrido un error al insertar el registro' . $e->getMessage() . ' ' . $e->getLine()], 422);
+            throw Utils::obtenerMensajeErrorLanzable($e);
         }
     }
 
-    public function show(PermisoEmpleado $permisoEmpleado)
+    public function show(PermisoEmpleado $permiso)
     {
-        $modelo = new PermisoEmpleadoResource($permisoEmpleado);
+        $modelo = new PermisoEmpleadoResource($permiso);
 
-        return response()->json(compact('modelo'), 200);
+        return response()->json(compact('modelo'));
     }
 
-    public function update(PermisoEmpleadoRequest $request, $permisoEmpleadoId)
+    /**
+     * @throws Exception
+     */
+    public function update(PermisoEmpleadoRequest $request, PermisoEmpleado $permiso)
     {
+        $autorizacion_id_old = $permiso->estado_permiso_id;
         $datos = $request->validated();
         $datos['estado_permiso_id'] = $request->safe()->only(['estado'])['estado'];
-        $permisoEmpleado = PermisoEmpleado::find($permisoEmpleadoId);
-        $permisoEmpleado->update($datos);
-        event(new PermisoEmpleadoEvent($permisoEmpleado));
-        if ($datos['estado_permiso_id'] == Autorizacion::APROBADO_ID) {
-            event(new PermisoNotificacionEvent($permisoEmpleado));
+        $permiso->update($datos);
+        if ($permiso->estado_permiso_id !== $autorizacion_id_old) {
+            event(new PermisoEmpleadoEvent($permiso));
+            if ($datos['estado_permiso_id'] == Autorizacion::APROBADO_ID) {
+                event(new PermisoNotificacionEvent($permiso));
+            }
         }
-        $modelo = new PermisoEmpleadoResource($permisoEmpleado);
+        $modelo = new PermisoEmpleadoResource($permiso);
         $mensaje = Utils::obtenerMensaje($this->entidad, 'update');
 
         return response()->json(compact('mensaje', 'modelo'));
     }
 
-    public function destroy($permisoEmpleadoId)
+    public function destroy(PermisoEmpleado $permiso)
     {
-        $permisoEmpleado = PermisoEmpleado::find($permisoEmpleadoId);
-        $permisoEmpleado->delete();
-        return $permisoEmpleado;
+        $permiso->delete();
+        $mensaje = Utils::obtenerMensaje($this->entidad, 'destroy');
+        return response()->json(compact('mensaje'));
     }
-    public function permisos_sin_recuperar(Request $request)
+
+    public function permisosSinRecuperar(Request $request)
     {
         $mes = Carbon::createFromFormat('m-Y', $request->mes)->format('Y-m');
 
         // Calcular el número total de días de permiso dentro del mes seleccionado usando funciones de agregación
-        $totalDiasPermiso = DB::table('permiso_empleados')
+        $total_dias_permiso = DB::table('permiso_empleados')
             ->selectRaw('SUM(DATEDIFF(fecha_hora_fin, fecha_hora_inicio) + 1) as total_dias_permiso')
             ->where('empleado_id', $request->empleado)
             ->whereRaw('DATE_FORMAT(fecha_hora_inicio, "%Y-%m") <= ?', [$mes])
@@ -159,6 +160,6 @@ class PermisoEmpleadoController extends Controller
             ->where('recupero', 0)
             ->value('total_dias_permiso');
 
-        return response()->json(compact('totalDiasPermiso'));
+        return response()->json(compact('total_dias_permiso'));
     }
 }
