@@ -16,7 +16,10 @@ use App\Models\RecursosHumanos\NominaPrestamos\RolPagoMes;
 use App\Models\User;
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
@@ -25,14 +28,16 @@ use Src\App\FondosRotativos\ReportePdfExcelService;
 use Src\App\RecursosHumanos\NominaPrestamos\NominaService;
 use Src\App\RecursosHumanos\NominaPrestamos\PrestamoService;
 use Src\Shared\Utils;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
+use Throwable;
 
 class RolPagoMesController extends Controller
 {
-    private $entidad = 'rol_pago';
-    private $reporteService;
-    private $nominaService;
-    private $prestamoService;
-    private $date;
+    private string $entidad = 'rol_pago';
+    private ReportePdfExcelService $reporteService;
+    private NominaService $nominaService;
+    private PrestamoService $prestamoService;
+    private string $date;
 
     public function __construct()
     {
@@ -44,7 +49,7 @@ class RolPagoMesController extends Controller
         $this->middleware('can:puede.crear.rol_pago_mes')->only('store');
     }
 
-    public function index(Request $request)
+    public function index()
     {
         $results = RolPagoMes::ignoreRequest(['campos'])->filter()->orderBy('id', 'desc')->get();
         $results = RolPagoMesResource::collection($results);
@@ -55,25 +60,26 @@ class RolPagoMesController extends Controller
      * La función de tienda en PHP se utiliza para crear un nuevo registro para el modelo RolPagoMes,
      * realizar comprobaciones de validación y manejar cualquier excepción que pueda ocurrir.
      *
-     * @param RolPagoMesRequest request El parámetro  es una instancia de la clase
+     * @param RolPagoMesRequest $request request El parámetro es una instancia de la clase
      * RolPagoMesRequest, que se utiliza para validar y recuperar los datos enviados en la solicitud HTTP.
      *
-     * @return una respuesta JSON que contiene las variables 'mensaje' y 'modelo'.
+     * @return JsonResponse respuesta JSON que contiene las variables 'mensaje' y 'modelo'.
+     * @throws Throwable
      */
     public function store(RolPagoMesRequest $request)
     {
         try {
             $datos = $request->validated();
             $existe_mes = RolPagoMes::where('mes', $request->mes)->where('es_quincena', '1')->get();
-            if ($request->es_quincena == false && count($existe_mes) == 0) {
+            if (!$request->es_quincena && count($existe_mes) == 0) {
                 throw ValidationException::withMessages([
                     '404' => ['Porfavor primero realice el Rol de Pagos de Quincena'],
                 ]);
             }
             DB::beginTransaction();
-            $rolPago = RolPagoMes::create($datos);
-            $modelo = new RolPagoMesResource($rolPago);
-            $this->tabla_roles($rolPago);
+            $rol = RolPagoMes::create($datos);
+            $modelo = new RolPagoMesResource($rol);
+            $this->tablaRoles($rol);
             DB::commit();
             $mensaje = Utils::obtenerMensaje($this->entidad, 'store');
             return response()->json(compact('mensaje', 'modelo'));
@@ -83,7 +89,6 @@ class RolPagoMesController extends Controller
             throw ValidationException::withMessages([
                 'Error al insertar registro' => [$e->getMessage() . $e->getLine()],
             ]);
-            return response()->json(['mensaje' => 'Ha ocurrido un error al insertar el registro de rol de pago' . $e->getMessage() . ' ' . $e->getLine()], 422);
         }
     }
     public function importarRolPago(RolPagoMesRequest $request)
@@ -91,7 +96,7 @@ class RolPagoMesController extends Controller
         try {
             DB::beginTransaction();
             $datos = $request->validated();
-            $rolPago = RolPagoMes::create($datos);
+            $rol = RolPagoMes::create($datos);
             $this->validate($request, [
                 'file' => 'required|mimes:xls,xlsx'
             ]);
@@ -100,7 +105,7 @@ class RolPagoMesController extends Controller
                     'file' => ['Debe seleccionar al menos un archivo.'],
                 ]);
             }
-            Excel::import(new RolPagoImport($request->mes, $rolPago), $request->file);
+            Excel::import(new RolPagoImport($request->mes, $rol), $request->file);
             DB::commit();
             return response()->json(['mensaje' => 'Subido exitosamente!']);
         } catch (Exception $e) {
@@ -110,44 +115,36 @@ class RolPagoMesController extends Controller
             ]);
         }
     }
+
     /**
      * La función "show" recupera un recurso "RolPagoMes" específico por su ID y lo devuelve como respuesta
      * JSON.
      *
-     * @param RolPagoMes rolPago Este parámetro es una instancia del modelo `RolPagoMes`. Se utiliza para
-     * recuperar el objeto `RolPagoMes` específico de la base de datos.
-     * @param rolPagoId El parámetro `` es el ID del objeto `RolPagoMes` que desea recuperar y
-     * mostrar.
-     *
-     * @return una respuesta JSON que contiene la variable "modelo", que es una instancia de la clase
+     * @param RolPagoMes $rol
+     * @return JsonResponse respuesta JSON que contiene la variable "modelo", que es una instancia de la clase
      * "RolPagoMesResource".
      */
-    public function show(RolPagoMes $rolPago,  $rolPagoId)
+    public function show(RolPagoMes $rol)
     {
-        $rolPago = RolPagoMes::find($rolPagoId);
-        $modelo = new RolPagoMesResource($rolPago);
-        return response()->json(compact('modelo'), 200);
+        $modelo = new RolPagoMesResource($rol);
+        return response()->json(compact('modelo'));
     }
 
     /**
      * La función actualiza un objeto RolPagoMes con los datos de solicitud proporcionados y devuelve una
      * respuesta JSON con un mensaje y el objeto actualizado.
      *
-     * @param Request request El parámetro  es una instancia de la clase Request, que representa la
-     * solicitud HTTP realizada al servidor. Contiene información sobre la solicitud, como el método de
-     * solicitud, encabezados y datos de entrada.
-     * @param rolPagoId El parámetro "rolPagoId" es el ID del objeto "RolPagoMes" que necesita ser
-     * actualizado.
-     *
-     * @return El código devuelve una respuesta JSON que contiene las variables "mensaje" y "modelo".
+     * @param Request $request
+     * @param RolPagoMes $rol
+     * @return JsonResponse código devuelve una respuesta JSON que contiene las variables "mensaje" y "modelo".
+     * @throws Throwable
      */
-    public function update(Request $request, $rolPagoId)
+    public function update(Request $request, RolPagoMes $rol)
     {
-        $rolPago = RolPagoMes::find($rolPagoId);
-        $rolPago->es_quincena = $request->es_quincena;
-        $rolPago->save();
-        $this->tabla_roles($rolPago, 'MODIFICAR');
-        $modelo = new RolPagoMesResource($rolPago);
+        $rol->es_quincena = $request->es_quincena;
+        $rol->save();
+        $this->tablaRoles($rol);
+        $modelo = new RolPagoMesResource($rol);
         $mensaje = Utils::obtenerMensaje($this->entidad, 'update');
         return response()->json(compact('mensaje', 'modelo'));
     }
@@ -155,28 +152,24 @@ class RolPagoMesController extends Controller
     /**
      * La función destruye un objeto RolPagoMes eliminándolo de la base de datos.
      *
-     * @param rolPagoId El parámetro `rolPagoId` es el ID del objeto `RolPagoMes` que debe eliminarse.
-     *
-     * @return La función `destroy` devuelve el objeto `RolPagoMes` eliminado.
+     * @param RolPagoMes $rol
+     * @return JsonResponse función `destroy` devuelve el objeto `RolPagoMes` eliminado.
      */
-    public function destroy($rolPagoId)
+    public function destroy(RolPagoMes $rol)
     {
-        $rolPago = RolPagoMes::find($rolPagoId);
-        $rolPago->delete();
-        return $rolPago;
+        $rol->delete();
+        $mensaje = Utils::obtenerMensaje($this->entidad, 'destroy');
+        return response()->json(compact('mensaje'));
     }
+
     /**
      * La función "imprimir_rol_pago_general" genera e imprime un informe general de nómina en PHP.
      *
-     * @param Request request El parámetro `` es una instancia de la clase
-     * `Illuminate\Http\Request`, que representa una solicitud HTTP. Contiene información sobre la
-     * solicitud, como el método de solicitud, encabezados y datos de entrada.
-     * @param rolPagoId El parámetro `` es el ID del rol_pago (nómina) para el cual desea generar
-     * un informe. Se utiliza para recuperar el rol_pago específico de la base de datos.
-     *
-     * @return el resultado del método `imprimir_reporte` del objeto ``.
+     * @param Request $request
+     * @param int $rolPagoId
+     * @return BinaryFileResponse|Response el resultado del método `imprimir_reporte` del objeto ``.
      */
-    public function imprimir_rol_pago_general(Request $request, $rolPagoId)
+    public function imprimirRolPagoGeneral(Request $request, int $rolPagoId)
     {
         try {
             $tipo = $request->tipo == 'xlsx' ? 'excel' : $request->tipo;
@@ -187,13 +180,13 @@ class RolPagoMesController extends Controller
                 ->get();
             $rol_pago =  RolPagoMes::where('id', $rolPagoId)->first();
             $es_quincena = $rol_pago->es_quincena;
-            $reportes = $this->generate_report_data($roles_pagos, $rol_pago->nombre);
+            $reportes = $this->generateReportData($roles_pagos, $rol_pago->nombre);
             $vista = $es_quincena ? 'recursos-humanos.rol_pago_quincena' : 'recursos-humanos.rol_pago_mes';
             // Log::channel('testing')->info('Log', ['191 - reportes', $es_quincena, $reportes]);
             $export_excel = new RolPagoMesExport($reportes, $es_quincena);
             $orientacion = $es_quincena ? 'portail' : 'landscape';
             $tipo_pagina = $es_quincena ? 'A4' : 'A2';
-            return $this->reporteService->imprimir_reporte($tipo,  $tipo_pagina, $orientacion, $reportes, $nombre_reporte, $vista, $export_excel);
+            return $this->reporteService->imprimirReporte($tipo,  $tipo_pagina, $orientacion, $reportes, $nombre_reporte, $vista, $export_excel);
         } catch (Exception $e) {
             Log::channel('testing')->info('Log', ['error', $e->getMessage(), $e->getLine()]);
             throw ValidationException::withMessages([
@@ -201,20 +194,19 @@ class RolPagoMesController extends Controller
             ]);
         }
     }
+
     /**
      * La función "enviarRoles" recupera una lista de roles de pago basada en un rolPagoId determinado,
      * recupera el empleado correspondiente para cada rol y luego envía el rol de pago al empleado
      * mediante nominaService.
      *
-     * @param rolPagoId El parámetro "rolPagoId" es el ID del rol_pago (rol de pago) que debe enviarse.
-     *
-     * @return una respuesta JSON con un mensaje indicando que la nómina ha sido enviada exitosamente.
+     * @param int $rolPagoId
+     * @return JsonResponse respuesta JSON con un mensaje indicando que la nómina ha sido enviada exitosamente.
      */
-    public function enviarRoles($rolPagoId)
+    public function enviarRoles(int $rolPagoId)
     {
-        $rolesPago = RolPago::where('rol_pago_id', $rolPagoId)->get();
-        $empleado = Empleado::where('id', 26)->first();
-        foreach ($rolesPago as $rol_pago) {
+        $roles = RolPago::where('rol_pago_id', $rolPagoId)->get();
+        foreach ($roles as $rol_pago) {
             $empleado = Empleado::where('id', $rol_pago->empleado_id)->first();
             $this->nominaService->enviar_rol_pago($rol_pago->id, $empleado);
         }
@@ -226,12 +218,10 @@ class RolPagoMesController extends Controller
      * La función crea un informe de pago de rol en efectivo en formato Excel para un rolPagoId
      * determinado.
      *
-     * @param rolPagoId El parámetro `rolPagoId` es el ID del rol_pago (nómina) para el que desea crear un
-     * rol_pago de efectivo.
-     *
-     * @return una descarga de un archivo Excel.
+     * @param int $rolPagoId
+     * @return BinaryFileResponse descarga de un archivo Excel.
      */
-    public function crear_cash_rol_pago($rolPagoId)
+    public function crearCashRolPago(int $rolPagoId)
     {
         $nombre_reporte = 'rol_pagos_general';
         $roles_pagos = RolPago::with(['egreso_rol_pago.descuento', 'ingreso_rol_pago.concepto_ingreso_info', 'rolPagoMes', 'egreso_rol_pago'])
@@ -248,17 +238,15 @@ class RolPagoMesController extends Controller
     }
 
     /**
-     * La función "imprimir_reporte_general" genera un informe en PHP, en base a la solicitud proporcionada
+     * La función "imprimir_reporte_general" genera un informe en PHP, en base, a la solicitud proporcionada
      * y rolPagoId, y lo exporta en el formato especificado (excel u otro).
      *
-     * @param Request request El parámetro  es una instancia de la clase Request, que se utiliza
-     * para recuperar datos de la solicitud HTTP.
-     * @param rolPagoId El parámetro `` es el ID del rolPago para el que desea generar un informe
-     * general.
-     *
-     * @return el resultado del método `imprimir_reporte` del objeto `reporteService`.
+     * @param Request $request
+     * @param int $rolPagoId
+     * @return Response|BinaryFileResponse resultado del método `imprimir_reporte` del objeto `reporteService`.
+     * @throws Exception
      */
-    public function imprimir_reporte_general(Request $request, $rolPagoId)
+    public function imprimirReporteGeneral(Request $request, int $rolPagoId)
     {
         $tipo = $request->tipo == 'xlsx' ? 'excel' : $request->tipo;
         $nombre_reporte = 'rol_pagos_general';
@@ -266,69 +254,69 @@ class RolPagoMesController extends Controller
         $sumatoria = RolPago::where('rol_pago_id', $rolPagoId)
             ->select(
                 DB::raw('SUM(ROUND(decimo_tercero, 2)) as decimo_tercero'),
-                DB::raw('SUM(ROUND(decimo_cuarto, 2))  as decimo_cuarto'),
-                DB::raw('SUM(ROUND(fondos_reserva, 2))  as fondos_reserva'),
-                DB::raw('SUM(ROUND(bonificacion, 2))  as bonificacion'),
-                DB::raw('SUM(ROUND(total_ingreso, 2))  as total_ingreso'),
-                DB::raw('SUM(ROUND(comisiones, 2))  as comisiones'),
-                DB::raw('SUM(ROUND(iess, 2))  as iess'),
-                DB::raw('SUM(ROUND(anticipo, 2))  as anticipo'),
-                DB::raw('SUM(ROUND(prestamo_quirorafario, 2))  as prestamo_quirorafario'),
-                DB::raw('SUM(ROUND(prestamo_hipotecario, 2))  as prestamo_hipotecario'),
-                DB::raw('SUM(ROUND(extension_conyugal, 2))  as extension_conyugal'),
-                DB::raw('SUM(ROUND(prestamo_empresarial, 2))  as prestamo_empresarial'),
-                DB::raw('SUM(ROUND(bono_recurente, 2))  as bono_recurente'),
-                DB::raw('SUM(ROUND(total_egreso, 2))  as total_egreso'),
-                DB::raw('SUM(ROUND(total, 2)) as total'),
+//                DB::raw('SUM(ROUND(decimo_cuarto, 2))  as decimo_cuarto'),
+//                DB::raw('SUM(ROUND(fondos_reserva, 2))  as fondos_reserva'),
+//                DB::raw('SUM(ROUND(bonificacion, 2))  as bonificacion'),
+//                DB::raw('SUM(ROUND(total_ingreso, 2))  as total_ingreso'),
+//                DB::raw('SUM(ROUND(comisiones, 2))  as comisiones'),
+//                DB::raw('SUM(ROUND(iess, 2))  as iess'),
+//                DB::raw('SUM(ROUND(anticipo, 2))  as anticipo'),
+//                DB::raw('SUM(ROUND(prestamo_quirorafario, 2))  as prestamo_quirorafario'),
+//                DB::raw('SUM(ROUND(prestamo_hipotecario, 2))  as prestamo_hipotecario'),
+//                DB::raw('SUM(ROUND(extension_conyugal, 2))  as extension_conyugal'),
+//                DB::raw('SUM(ROUND(prestamo_empresarial, 2))  as prestamo_empresarial'),
+//                DB::raw('SUM(ROUND(bono_recurente, 2))  as bono_recurente'),
+//                DB::raw('SUM(ROUND(total_egreso, 2))  as total_egreso'),
+//                DB::raw('SUM(ROUND(total, 2)) as total'),
             )
             ->first();
         $results = RolPago::empaquetarListado($roles_de_pago);
-        $column_names_ingresos = $this->extract_column_names($results, 'ingresos', 'concepto_ingreso_info', 'nombre');
+        $column_names_ingresos = $this->extractColumnNames($results, 'ingresos', 'concepto_ingreso_info', 'nombre');
         $column_names_ingresos =  array_unique($column_names_ingresos['ingresos']);
-        $column_names_egresos = $this->extract_column_names($results, 'egresos', 'descuento', 'nombre');
+        $column_names_egresos = $this->extractColumnNames($results, 'egresos', 'descuento', 'nombre');
         $column_names_egresos = array_unique($column_names_egresos['egresos']);
-        $colum_ingreso_value = $this->colum_values($results,  $column_names_ingresos, 'ingresos', 'concepto_ingreso_info');
-        $colum_egreso_value = $this->colum_values($results, $column_names_egresos, 'egresos', 'descuento');
-        $rolPago = RolPagoMes::where('id', $rolPagoId)->first();
-        $reportes = ['reporte' => $sumatoria, 'rolPago' => $rolPago, 'ingresos' => $this->sumatoria_llaves($colum_ingreso_value), 'egresos' => $this->sumatoria_llaves($colum_egreso_value)];
+        $colum_ingreso_value = $this->columValues($results,  $column_names_ingresos, 'ingresos', 'concepto_ingreso_info');
+        $colum_egreso_value = $this->columValues($results, $column_names_egresos, 'egresos', 'descuento');
+        $rol = RolPagoMes::where('id', $rolPagoId)->first();
+        $reportes = ['reporte' => $sumatoria, 'rolPago' => $rol, 'ingresos' => $this->sumatoriaLlaves($colum_ingreso_value), 'egresos' => $this->sumatoriaLlaves($colum_egreso_value)];
         $vista = 'recursos-humanos.reporte_general';
         $export_excel = new RolPagoGeneralExport($reportes);
-        return $this->reporteService->imprimir_reporte($tipo, 'A4', 'landscape', $reportes, $nombre_reporte, $vista, $export_excel);
+        return $this->reporteService->imprimirReporte($tipo, 'A4', 'landscape', $reportes, $nombre_reporte, $vista, $export_excel);
     }
+
     /**
      * La función "generate_report_data" en PHP genera un informe procesando una serie de roles de pago y
      * extrayendo datos relevantes para su visualización.
      *
-     * @param roles_pagos Una serie de datos de pago de roles.
-     * @param nombre El parámetro "nombre" es una cadena que representa el nombre del informe.
-     *
-     * @return una matriz con las siguientes claves:
+     * @param array|Collection $roles_pagos $roles_pagos Una serie de datos de pago de roles.
+     * @param string $nombre
+     * @return array matriz con las siguientes claves:
      */
-    private function generate_report_data($roles_pagos, $nombre)
+    private function generateReportData(array|Collection $roles_pagos,string $nombre)
     {
-        $es_quincena = RolPagoMes::where('mes', $roles_pagos[0]->mes)->where('es_quincena', '1')->first() != null ? true : false;
+        $es_quincena = RolPagoMes::where('mes', $roles_pagos[0]->mes)->where('es_quincena', '1')->first() != null;
         $periodo = $this->obtenerPeriodo($roles_pagos[0]->mes, $es_quincena);
         $departamento_gerencia = Departamento::where("nombre", Departamento::DEPARTAMENTO_GERENCIA)->first();
-        $responsable_gerencia = $departamento_gerencia ?  $departamento_gerencia?->responsable : Empleado::find(117);
+        $responsable_gerencia = $departamento_gerencia ?  $departamento_gerencia->responsable : Empleado::find(117);
         $creador_rol_pago = User::role(User::ROL_RECURSOS_HUMANOS)->permission('puede.elaborar.rol_pago')->first()->empleado;
         $results = RolPago::empaquetarListado($roles_pagos);
         $results = collect($results)->map(function ($elemento, $index) {
             $elemento['item'] = $index + 1;
             return $elemento;
         })->all();
-        $column_names_egresos = $this->extract_column_names($results, 'egresos', 'descuento', 'abreviatura');
-        $column_names_ingresos = $this->extract_column_names($results, 'ingresos', 'concepto_ingreso_info', 'abreviatura');
+        $column_names_egresos = $this->extractColumnNames($results, 'egresos', 'descuento', 'abreviatura');
+        $column_names_ingresos = $this->extractColumnNames($results, 'ingresos', 'concepto_ingreso_info', 'abreviatura');
         $columnas_ingresos =  array_unique($column_names_ingresos['ingresos']);
-        $colum_ingreso_value = $this->colum_values($results, $columnas_ingresos, 'ingresos', 'concepto_ingreso_info', true);
+        $colum_ingreso_value = $this->columValues($results, $columnas_ingresos, 'ingresos', 'concepto_ingreso_info', true);
         $columnas_egresos = array_unique($column_names_egresos['egresos']);
-        $colum_egreso_value = $this->colum_values($results, $columnas_egresos, 'egresos', 'descuento', true);
-        $maxColumEgresosValue = count($columnas_egresos);
-        $maxColumIngresosValue = count($columnas_ingresos);
+        $colum_egreso_value = $this->columValues($results, $columnas_egresos, 'egresos', 'descuento', true);
+        $max_column_egresos_value = count($columnas_egresos);
+        $max_column_ingresos_value = count($columnas_ingresos);
 
 
 
         // Calculate the sum of specific columns from the main data array
-        $sumColumns = [
+        $sum_columns = [
             'salario' => 0,
             'sueldo' => 0,
             'decimo_tercero' => 0,
@@ -350,53 +338,51 @@ class RolPagoMesController extends Controller
 
         // Itera a través del array $results y suma los valores en las columnas
         foreach ($results as $item) {
-            $sumColumns['salario'] += $item['salario'];
-            $sumColumns['sueldo'] += $item['sueldo'];
-            $sumColumns['decimo_tercero'] += $item['decimo_tercero'];
-            $sumColumns['decimo_cuarto'] += $item['decimo_cuarto'];
-            $sumColumns['fondos_reserva'] += $item['fondos_reserva'];
-            $sumColumns['iess'] += $item['iess'];
-            $sumColumns['anticipo'] += $item['anticipo'];
-            $sumColumns['bonificacion'] += $item['bonificacion'];
-            $sumColumns['bono_recurente'] += $item['bono_recurente'];
-            $sumColumns['total_ingreso'] += $item['total_ingreso'];
-            $sumColumns['prestamo_quirorafario'] += $item['prestamo_quirorafario'];
-            $sumColumns['prestamo_hipotecario'] += $item['prestamo_hipotecario'];
-            $sumColumns['extension_conyugal'] += $item['extension_conyugal'];
-            $sumColumns['prestamo_empresarial'] += $item['prestamo_empresarial'];
-            $sumColumns['supa'] += $item['supa'];
-            $sumColumns['total_egreso'] += $item['total_egreso'];
-            $sumColumns['total'] += $item['total'];
+            $sum_columns['salario'] += $item['salario'];
+            $sum_columns['sueldo'] += $item['sueldo'];
+            $sum_columns['decimo_tercero'] += $item['decimo_tercero'];
+            $sum_columns['decimo_cuarto'] += $item['decimo_cuarto'];
+            $sum_columns['fondos_reserva'] += $item['fondos_reserva'];
+            $sum_columns['iess'] += $item['iess'];
+            $sum_columns['anticipo'] += $item['anticipo'];
+            $sum_columns['bonificacion'] += $item['bonificacion'];
+            $sum_columns['bono_recurente'] += $item['bono_recurente'];
+            $sum_columns['total_ingreso'] += $item['total_ingreso'];
+            $sum_columns['prestamo_quirorafario'] += $item['prestamo_quirorafario'];
+            $sum_columns['prestamo_hipotecario'] += $item['prestamo_hipotecario'];
+            $sum_columns['extension_conyugal'] += $item['extension_conyugal'];
+            $sum_columns['prestamo_empresarial'] += $item['prestamo_empresarial'];
+            $sum_columns['supa'] += $item['supa'];
+            $sum_columns['total_egreso'] += $item['total_egreso'];
+            $sum_columns['total'] += $item['total'];
         }
         // El resultado deseado se encuentra ahora en el array $sumColumns
         return [
             'roles_pago' => $results,
             'periodo' => $periodo,
-            'cantidad_columna_ingresos' => $maxColumIngresosValue,
-            'cantidad_columna_egresos' => $maxColumEgresosValue,
+            'cantidad_columna_ingresos' => $max_column_ingresos_value,
+            'cantidad_columna_egresos' => $max_column_egresos_value,
             'colum_ingreso_value' => $colum_ingreso_value,
             'colum_egreso_value' => $colum_egreso_value,
             'columnas_ingresos' => $columnas_ingresos,
             'columnas_egresos' =>  $columnas_egresos,
-            'sumatoria' => $sumColumns,
+            'sumatoria' => $sum_columns,
             'nombre' => $nombre,
             'creador_rol_pago' => $creador_rol_pago,
             'aprueba_rol_pago' => $responsable_gerencia,
-            'sumatoria_ingresos' => $this->calculate_column_sum($results, $maxColumIngresosValue, 'ingresos_cantidad_columna', 'ingresos'),
-            'sumatoria_egresos' => $this->sumatoria_llaves($colum_egreso_value),
+            'sumatoria_ingresos' => $this->calculateColumnSum($results, $max_column_ingresos_value),
+            'sumatoria_egresos' => $this->sumatoriaLlaves($colum_egreso_value),
         ];
     }
+
     /**
      * La función "sumatoria_keys" calcula la suma de valores de cada clave en una matriz asociativa.
      *
-     * @param data El parámetro "datos" es una matriz que contiene pares clave-valor. Cada clave representa
-     * una categoría o grupo y el valor correspondiente es una matriz de entradas. Cada entrada tiene un
-     * campo "valor", que representa un valor numérico.
-     *
-     * @return un array asociativo llamado . Esta matriz contiene la suma del campo
-     * "valor" para cada clave en la matriz .
+     * @param array $data
+     * @return array array asociativo llamado `sumatoria_por_llaves`. Esta matriz contiene la suma del campo
+     * "valor" para cada clave en la matriz `sumatoria_por_llaves`.
      */
-    private function sumatoria_llaves($data)
+    private function sumatoriaLlaves(array $data)
     {
         // Inicializa un arreglo asociativo para almacenar la sumatoria por llaves
         $sumatoria_por_llaves = [];
@@ -410,52 +396,48 @@ class RolPagoMesController extends Controller
         }
         return $sumatoria_por_llaves;
     }
+
     /**
      * La función `colum_values` toma una matriz de datos, nombres de columnas y dos claves, y agrupa los
      * datos por un nombre de columna específico.
      *
-     * @param data El parámetro "datos" es una matriz de objetos. Cada objeto representa una fila de
-     * datos.
-     * @param column_name Una matriz que contiene los nombres de las columnas.
-     * @param key1 El parámetro `` se utiliza para acceder a una clave específica en la matriz
-     * ``. Probablemente se use para recuperar una submatriz u objeto anidado dentro de cada elemento
-     * de la matriz ``.
-     * @param key2 El parámetro  se utiliza para acceder a la propiedad "nombre" del elemento en la
-     * matriz []. Se usa para determinar el valor de , que luego se usa para agrupar
-     * los objetos en el arreglo .
-     *
-     * @return una matriz llamada .
+     * @param array $data
+     * @param array $column_name
+     * @param string $key1
+     * @param string $key2
+     * @param bool $abreviatura
+     * @return array matriz llamada .
      */
-    private function colum_values($data, $column_name, $key1, $key2, $abreviatura = false)
+    private function columValues(array $data, array $column_name, string $key1, string $key2, bool $abreviatura = false)
     {
         // Creamos un arreglo para almacenar los objetos agrupados por descuento_id
-        $groupedData = [];
+        $grouped_data = [];
         foreach ($data as $item) {
             // Recorremos el arreglo original y agrupamos los objetos por descuento_id
             foreach ($item[$key1] as $item) {
-                $descuentoId = $abreviatura ? $item[$key2]->abreviatura : $item[$key2]->nombre;
-                if (!isset($groupedData[$descuentoId])) {
-                    $groupedData[$descuentoId] = [];
+                $descuento_id = $abreviatura ? $item[$key2]->abreviatura : $item[$key2]->nombre;
+                if (!isset($grouped_data[$descuento_id])) {
+                    $grouped_data[$descuento_id] = [];
                 }
                 foreach ($column_name as $name) {
-                    if ($name != $descuentoId) {
-                        $groupedData[$name][] = ['id' => $item['id_rol_pago'], 'valor' => 0];
+                    if ($name != $descuento_id) {
+                        $grouped_data[$name][] = ['id' => $item['id_rol_pago'], 'valor' => 0];
                     }
-                    if ($name == $descuentoId) {
-                        $groupedData[$descuentoId][] = ['id' => $item['id_rol_pago'], 'valor' => $item['monto']];
+                    if ($name == $descuento_id) {
+                        $grouped_data[$descuento_id][] = ['id' => $item['id_rol_pago'], 'valor' => $item['monto']];
                     }
                 }
             }
         }
-        return $this->eliminar_duplicados($groupedData, 'id');
+        return $this->eliminarDuplicados($grouped_data);
     }
-    public function eliminar_duplicados(array $datos, $llave_buscar): array
+    public function eliminarDuplicados(array $datos): array
     {
         $array_sin_duplicados = [];
         foreach ($datos as $clave => $objeto) {
             $array_sin_duplicados[$clave] = [];
             $serial_ant = null;
-            if ($this->verificar_valor_cero_primera_posicion($objeto)) {
+            if ($this->verificarValorCeroPrimeraPosicion($objeto)) {
                 rsort($objeto);
             }
             foreach ($objeto as $objeto_actual) {
@@ -469,7 +451,7 @@ class RolPagoMesController extends Controller
 
         return $array_sin_duplicados;
     }
-    function verificar_valor_cero_primera_posicion($array)
+    function verificarValorCeroPrimeraPosicion($array)
     {
         if ($array[0]['valor'] == 0) {
             return true;
@@ -478,24 +460,18 @@ class RolPagoMesController extends Controller
     }
 
 
-
     /**
      * La función extrae nombres de columnas de una matriz multidimensional basada en claves especificadas
      * y un nombre de columna.
      *
-     * @param results Una serie de resultados de una consulta de base de datos. Cada elemento de la matriz
-     * representa una fila de datos.
-     * @param key1 El parámetro `key1` es una cadena que representa la clave utilizada para acceder a una
-     * matriz dentro de la matriz ``.
-     * @param key2 El parámetro `key2` se utiliza para acceder a una clave específica dentro de la matriz
-     * ``.
-     * @param columnName El parámetro `columnName` es una cadena que representa el nombre de la columna que
-     * desea extraer de las matrices anidadas en la matriz ``.
-     *
-     * @return una serie de nombres de columnas. La matriz tiene dos claves, 'egresos' e 'ingresos', cada
+     * @param array $results
+     * @param string $key1
+     * @param string $key2
+     * @param string $columnName
+     * @return array|array[] serie de nombres de columnas. La matriz tiene dos claves, 'egresos' e 'ingresos', cada
      * una de las cuales contiene una matriz de nombres de columnas.
      */
-    private function extract_column_names($results, $key1, $key2, $columnName)
+    private function extractColumnNames(array $results,string $key1,string $key2,string $columnName)
     {
         $column_names = ['egresos' => [], 'ingresos' => []];
         foreach ($results as $item) {
@@ -507,26 +483,21 @@ class RolPagoMesController extends Controller
         }
         return $column_names;
     }
+
     /**
      * La función calcula la suma de valores en una columna específica de una matriz, según ciertas
      * condiciones.
      *
-     * @param data Una matriz de datos que contiene múltiples elementos.
-     * @param maximo La variable `` representa el número máximo de iteraciones o elementos en un
-     * bucle o matriz. Se utiliza en el bucle `for` para iterar un número específico de veces.
-     * @param key_cantidad El parámetro `` se utiliza como clave para acceder al valor de
-     * cantidad en la matriz ``. Se utiliza para comprobar si la cantidad es mayor que 0 o igual a
-     * 0 para poder realizar ciertos cálculos.
-     * @param key1 El parámetro `` se utiliza como clave para acceder a una matriz anidada dentro
-     * de la matriz ``. Se utiliza en el bucle foreach para iterar sobre los elementos de la matriz
-     * anidada.
-     *
-     * @return una matriz que contiene la suma de valores para cada clave única en los datos de entrada.
+     * @param $data
+     * @param $maximo
+     * @return array matriz que contiene la suma de valores para cada clave única en los datos de entrada.
      */
-    private function calculate_column_sum($data, $maximo, $key_cantidad, $key1)
+    private function calculateColumnSum($data, $maximo)
     {
-        $totalMontoIngresos = array_map(
-            function ($item) use ($maximo, $key_cantidad, $key1, $data) {
+        $total_monto_ingresos = array_map(
+            function ($item) use ($maximo, $data) {
+                $key1 = 'ingresos';
+                $key_cantidad = 'ingresos_cantidad_columna';
                 $monto = array();
                 if ($item[$key_cantidad] > 0) {
                     foreach ($item[$key1] as $subitem) {
@@ -543,7 +514,7 @@ class RolPagoMesController extends Controller
             $data
         );
         $resultados = [];
-        foreach ($totalMontoIngresos as $elemento) {
+        foreach ($total_monto_ingresos as $elemento) {
             // Verifica si el elemento es un arreglo asociativo (objeto JSON)
             if (is_array($elemento) && count($elemento) > 0) {
                 foreach ($elemento as $clave => $valor) {
@@ -568,17 +539,16 @@ class RolPagoMesController extends Controller
         }
         return $arreglo;
     }
+
     /**
      * La función "tabla_roles" calcula e inserta datos de nómina para empleados activos en función de
      * varios factores, como salario, asignaciones, deducciones y préstamos.
      *
-     * @param RolPagoMes rol El parámetro `` es una instancia de la clase `RolPagoMes`. Representa
-     * un mes de nómina y contiene información como el mes y si es una nómina quincenal o no.
-     *
-     * @return La función no devuelve nada. Está insertando datos en la tabla de la base de datos
-     * "RolPago" usando el método `insert()`.
+     * @param RolPagoMes $rol
+     * @return void
+     * @throws Throwable
      */
-    private function tabla_roles(RolPagoMes $rol)
+    private function tablaRoles(RolPagoMes $rol)
     {
         Log::channel('testing')->info('Log', ['rol', $rol]);
         try {
@@ -590,7 +560,7 @@ class RolPagoMesController extends Controller
                 ->where('esta_en_rol_pago', true)
                 ->where('salario', '!=', 0)
                 ->where('fecha_vinculacion', '<', $ultimo_dia_mes)
-                ->orderBy('apellidos', 'asc')->get();
+                ->orderBy('apellidos')->get();
             $this->nominaService->setMes($mes);
             $this->prestamoService->setMes($mes);
             $roles_pago = [];
@@ -598,8 +568,8 @@ class RolPagoMesController extends Controller
                 $this->nominaService->setEmpleado($empleado->id);
                 $this->prestamoService->setEmpleado($empleado->id);
                 // Calcular el número total de días de permiso dentro del mes seleccionado usando funciones de agregación
-                $diasTranscurridos = $rol->es_quincena ? 15 : 30;
-                $dias = $this->nominaService->calcularDias($diasTranscurridos);
+                $dias_transcurridos = $rol->es_quincena ? 15 : 30;
+                $dias = $this->nominaService->calcularDias($dias_transcurridos);
                 //$salario = $this->nominaService->calcularSalario();
                 $salario = $empleado->salario;
                 $sueldo =  $this->nominaService->calcularSueldo($dias, $rol->es_quincena);
@@ -648,20 +618,23 @@ class RolPagoMesController extends Controller
             ]);
         }
     }
+
+
     /**
      * La función "actualizar_tabla_roles" calcula y actualiza la tabla roles_pago de los empleados en
      * función de diversos cálculos de salarios y deducciones.
      *
-     * @param RolPagoMes rol El parámetro `` es una instancia de la clase `RolPagoMes`.
+     * @param RolPagoMes $rol
+     * @throws Throwable
      */
-    public function agregar_nuevos_empleados(RolPagoMes $rol)
+    public function agregarNuevosEmpleados(RolPagoMes $rol)
     {
         try {
             // $rol_pago = RolPagoMes::find($rolPagoId);
             $mes_rol = Carbon::createFromFormat('m-Y', $rol->mes)->format('Y-m');
             $final_mes = new Carbon($mes_rol);
             $ultimo_dia_mes = $final_mes->endOfMonth();
-            $empleadosSinRolPago = Empleado::where('id', '>', 2)
+            $empleados_sin_rol_pago = Empleado::where('id', '>', 2)
                 ->where('estado', true)
                 ->where('esta_en_rol_pago', true)
                 ->where('fecha_vinculacion', '<', $ultimo_dia_mes)
@@ -674,11 +647,11 @@ class RolPagoMesController extends Controller
             $this->nominaService->setMes($mes);
             $this->prestamoService->setMes($mes);
             $roles_pago = [];
-            foreach ($empleadosSinRolPago as $empleado) {
+            foreach ($empleados_sin_rol_pago as $empleado) {
                 $this->nominaService->setEmpleado($empleado->id);
                 $this->prestamoService->setEmpleado($empleado->id);
-                $diasTranscurridos = $rol->es_quincena ? 15 : 30;
-                $dias = $this->nominaService->calcularDias($diasTranscurridos);
+                $dias_transcurridos = $rol->es_quincena ? 15 : 30;
+                $dias = $this->nominaService->calcularDias($dias_transcurridos);
                 $salario = $this->nominaService->calcularSalario();
                 // Calcular el número total de días de permiso dentro del mes seleccionado usando funciones de agregación
                 $sueldo =  $this->nominaService->calcularSueldo($dias, $rol->es_quincena);
@@ -732,18 +705,23 @@ class RolPagoMesController extends Controller
     public function verificarTodasRolesFinalizadas(Request $request)
     {
         $rol_pago = RolPagoMes::find($request['rol_pago_id']);
-        $totalSubrol_pagosNoFinalizadas = $rol_pago->rolPago()->whereIn('estado', [RolPago::EJECUTANDO, RolPago::REALIZADO])->count();
-        $estan_finalizadas = $totalSubrol_pagosNoFinalizadas == 0;
+        $total_subrol_pagos_no_finalizadas = $rol_pago->rolPago()->whereIn('estado', [RolPago::EJECUTANDO, RolPago::REALIZADO])->count();
+        $estan_finalizadas = $total_subrol_pagos_no_finalizadas == 0;
         return response()->json(compact('estan_finalizadas'));
     }
-    public function actualizar_tabla_roles(RolPagoMes $rol_mes)
+
+    /**
+     * @param RolPagoMes $rol_mes
+     * @return void
+     */
+    public function actualizarTablaRoles(RolPagoMes $rol_mes)
     {
         try {
             $mes = Carbon::createFromFormat('m-Y', $rol_mes->mes)->format('Y-m');
             $this->nominaService->setMes($mes);
             $this->prestamoService->setMes($mes);
             $roles_pago =  RolPago::where('rol_pago_id', $rol_mes->id)->get();
-            foreach ($roles_pago as $key => $rol_pago) {
+            foreach ($roles_pago as $rol_pago) {
                 $this->nominaService->setEmpleado($rol_pago->empleado_id);
                 $this->prestamoService->setEmpleado($rol_pago->empleado_id);
                 $this->nominaService->setRolPago($rol_mes);
@@ -798,17 +776,18 @@ class RolPagoMesController extends Controller
             throw ValidationException::withMessages([
                 'Error al refrescar rol pago por empleado' => [$ex->getMessage()],
             ]);
+        } catch (Throwable $e) {
+            throw Utils::obtenerMensajeErrorLanzable($e);
         }
     }
-    public function refrescar_rol_pago($rolPagoId)
+    public function refrescarRolPago(RolPagoMes $rol)
     {
-        $rol_pago = RolPagoMes::find($rolPagoId);
         // $this->agregar_nuevos_empleados($rol_pago);
-        $this->actualizar_tabla_roles($rol_pago);
+        $this->actualizarTablaRoles($rol);
         $mensaje = "Rol de pago Actualizado Exitosamente";
         return response()->json(compact('mensaje'));
     }
-    public function FinalizarRolPago(Request $request)
+    public function finalizarRolPago(Request $request)
     {
         $rol_pago = RolPagoMes::find($request['rol_pago_id']);
         $rol_pago->finalizado = true;
@@ -819,7 +798,7 @@ class RolPagoMesController extends Controller
             // Divide la fecha en mes y año
             list($month, $year) = explode('-', $mes);
             // Crea un objeto Carbon para representar la fecha en Laravel
-            $date = \Carbon\Carbon::createFromDate($year, $month, 1);
+            $date = Carbon::createFromDate($year, $month, 1);
             // Formatea la fecha en el formato deseado
             $mes = $date->format('Y-m');
             $this->prestamoService->setMes($mes);
