@@ -6,46 +6,45 @@ namespace App\Http\Controllers;
 
 use App\Events\TransaccionEgresoEvent;
 use App\Exports\TransaccionBodegaEgresoExport;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\DB;
-use Barryvdh\DomPDF\Facade\Pdf;
-use Illuminate\Http\Request;
-use Src\Shared\Utils;
-use Exception;
-
-// Modelos
-use App\Models\MaterialEmpleadoTarea;
-use App\Models\DetalleProducto;
-use App\Models\TipoTransaccion;
-use App\Models\Inventario;
-use App\Models\Empleado;
-use App\Models\Motivo;
-use App\Models\TransaccionBodega;
-use App\Models\User;
-
-// Logica
-use App\Http\Resources\TransaccionBodegaResource;
 use App\Http\Requests\TransaccionBodegaRequest;
 use App\Http\Resources\ClienteResource;
+use App\Http\Resources\TransaccionBodegaResource;
 use App\Models\Cliente;
 use App\Models\Comprobante;
 use App\Models\ConfiguracionGeneral;
+use App\Models\DetalleProducto;
 use App\Models\DetalleProductoTransaccion;
+use App\Models\Empleado;
 use App\Models\EstadoTransaccion;
+use App\Models\Inventario;
 use App\Models\MaterialEmpleado;
+use App\Models\MaterialEmpleadoTarea;
 use App\Models\Pedido;
+use App\Models\TransaccionBodega;
+use App\Models\User;
+use Barryvdh\DomPDF\Facade\Pdf;
+use Exception;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 use Maatwebsite\Excel\Facades\Excel;
 use Src\App\Tareas\ProductoEmpleadoService;
 use Src\App\Tareas\ProductoTareaEmpleadoService;
 use Src\App\TransaccionBodegaEgresoService;
 use Src\Config\Autorizaciones;
+use Src\Shared\Utils;
+use Throwable;
+
+// Modelos
+
+// Logica
 
 class TransaccionBodegaEgresoController extends Controller
 {
-    private $entidad = 'Transacción';
-    private $servicio;
-    private $productosEmpleadoService;
-    private $productosTareaEmpleadoService;
+    private string $entidad = 'Transacción';
+    private TransaccionBodegaEgresoService $servicio;
+    private ProductoEmpleadoService $productosEmpleadoService;
+    private ProductoTareaEmpleadoService $productosTareaEmpleadoService;
 
     public function __construct()
     {
@@ -60,14 +59,14 @@ class TransaccionBodegaEgresoController extends Controller
 
 
     // Stock personal: solo materiales excepto bobinas
-    public function obtenerMaterialesEmpleado(Request $request)
+    public function obtenerMaterialesEmpleado()
     {
         $results = $this->productosEmpleadoService->obtenerProductos();
         return response()->json(compact('results'));
     }
 
     // Stock personal: materiales y bobinas material para tarea no borrar
-    public function obtenerMaterialesEmpleadoTarea(Request $request)
+    public function obtenerMaterialesEmpleadoTarea()
     {
         $results = $this->productosTareaEmpleadoService->obtenerProductos();
         return response()->json(compact('results'));
@@ -75,7 +74,6 @@ class TransaccionBodegaEgresoController extends Controller
 
     public function obtenerMaterialesEmpleadoConsolidado(Request $request)
     {
-        $results = [];
         try {
             if (!$request->exists('cliente_id')) $request->merge(['cliente_id' => null]);
             $request->validate([
@@ -100,7 +98,7 @@ class TransaccionBodegaEgresoController extends Controller
             });
 
             return response()->json(compact('results'));
-        } catch (\Throwable $th) {
+        } catch (Throwable $th) {
             $mensaje = $th->getMessage() . '. ' . $th->getLine();
             return response()->json(compact('mensaje'));
         }
@@ -111,9 +109,9 @@ class TransaccionBodegaEgresoController extends Controller
      */
     public function index(Request $request)
     {
-        $results = $this->servicio->listar($request);
-        $results = TransaccionBodegaResource::collection($results);
-        return response()->json(compact('results'));
+        $results = $this->servicio->listar($request, $request->paginate);
+        return TransaccionBodegaResource::collection($results);
+//        return response()->json(compact('results'));
     }
 
 
@@ -127,25 +125,25 @@ class TransaccionBodegaEgresoController extends Controller
             $datos = $request->validated();
             DB::beginTransaction();
             // $datos['tipo_id'] = $request->safe()->only(['tipo'])['tipo'];
-            if ($request->pedido) $datos['pedido_id'] = $request->safe()->only(['pedido'])['pedido'];
-            if ($request->transferencia) $datos['transferencia_id'] = $request->safe()->only(['transferencia'])['transferencia'];
-            $datos['motivo_id'] = $request->safe()->only(['motivo'])['motivo'];
-            $datos['solicitante_id'] = $request->safe()->only(['solicitante'])['solicitante'];
-            $datos['sucursal_id'] = $request->safe()->only(['sucursal'])['sucursal'];
-            $datos['per_autoriza_id'] = $request->safe()->only(['per_autoriza'])['per_autoriza'];
-            $datos['per_retira_id'] = $request->safe()->only(['per_retira'])['per_retira'];
-            $datos['cliente_id'] = $request->safe()->only(['cliente'])['cliente'];
-            $datos['responsable_id'] = $request->safe()->only(['responsable'])['responsable'];
-            $datos['per_retira_id'] = $request->safe()->only(['per_retira'])['per_retira'];
-            if ($request->proyecto) $datos['proyecto_id'] = $request->safe()->only(['proyecto'])['proyecto'];
-            if ($request->etapa) $datos['etapa_id'] = $request->safe()->only(['etapa'])['etapa'];
-            if ($request->tarea) $datos['tarea_id'] = $request->safe()->only(['tarea'])['tarea'];
-            if ($request->subtarea) $datos['subtarea_id'] = $request->safe()->only(['subtarea'])['subtarea'];
-            if ($request->per_atiende) $datos['per_atiende_id'] = $request->safe()->only(['per_atiende'])['per_atiende'];
+//            if ($request->pedido) $datos['pedido_id'] = $request->safe()->only(['pedido'])['pedido'];
+//            if ($request->transferencia) $datos['transferencia_id'] = $request->safe()->only(['transferencia'])['transferencia'];
+//            $datos['motivo_id'] = $request->safe()->only(['motivo'])['motivo'];
+//            $datos['solicitante_id'] = $request->safe()->only(['solicitante'])['solicitante'];
+//            $datos['sucursal_id'] = $request->safe()->only(['sucursal'])['sucursal'];
+//            $datos['per_autoriza_id'] = $request->safe()->only(['per_autoriza'])['per_autoriza'];
+//            $datos['per_retira_id'] = $request->safe()->only(['per_retira'])['per_retira'];
+//            $datos['cliente_id'] = $request->safe()->only(['cliente'])['cliente'];
+//            $datos['responsable_id'] = $request->safe()->only(['responsable'])['responsable'];
+//            $datos['per_retira_id'] = $request->safe()->only(['per_retira'])['per_retira'];
+//            if ($request->proyecto) $datos['proyecto_id'] = $request->safe()->only(['proyecto'])['proyecto'];
+//            if ($request->etapa) $datos['etapa_id'] = $request->safe()->only(['etapa'])['etapa'];
+//            if ($request->tarea) $datos['tarea_id'] = $request->safe()->only(['tarea'])['tarea'];
+//            if ($request->subtarea) $datos['subtarea_id'] = $request->safe()->only(['subtarea'])['subtarea'];
+//            if ($request->per_atiende) $datos['per_atiende_id'] = $request->safe()->only(['per_atiende'])['per_atiende'];
 
             //datos de las relaciones muchos a muchos
-            $datos['autorizacion_id'] = $request->safe()->only(['autorizacion'])['autorizacion'];
-            $datos['estado_id'] = $request->safe()->only(['estado'])['estado'];
+//            $datos['autorizacion_id'] = $request->safe()->only(['autorizacion'])['autorizacion'];
+//            $datos['estado_id'] = $request->safe()->only(['estado'])['estado'];
 
 
             //Creacion de la transaccion
@@ -154,13 +152,13 @@ class TransaccionBodegaEgresoController extends Controller
             //Guardar los productos seleccionados
             foreach ($request->listadoProductosTransaccion as $listado) {
                 // $itemInventario = Inventario::where('detalle_id', $listado['detalle'])->first();
-                $itemInventario = Inventario::find($listado['id']);
-                $transaccion->items()->attach($itemInventario->id, ['cantidad_inicial' => $listado['cantidad']]);
+                $item_inventario = Inventario::find($listado['id']);
+                $transaccion->items()->attach($item_inventario->id, ['cantidad_inicial' => $listado['cantidad']]);
                 //Actualizamos el estado del item de inventario
-                TransaccionBodega::desactivarDetalle($itemInventario->detalle_id);
+                TransaccionBodega::desactivarDetalle($item_inventario->detalle_id);
                 // Actualizamos la cantidad en inventario
-                $itemInventario->cantidad -= $listado['cantidad'];
-                $itemInventario->save();
+                $item_inventario->cantidad -= $listado['cantidad'];
+                $item_inventario->save();
             }
 
             //Si hay pedido, actualizamos su estado.
@@ -174,14 +172,14 @@ class TransaccionBodegaEgresoController extends Controller
 
             $modelo = new TransaccionBodegaResource($transaccion);
 
-            //verificamos si es un egreso por transferencia, en ese caso habría responsable de los materiales pero no se crea comprobante,
+            //verificamos si es un egreso por transferencia, en ese caso habría responsable de los materiales, pero no se crea comprobante,
             if (!$transaccion->transferencia_id) {
-                $noGeneraComprobante = TransaccionBodega::verificarMotivosEgreso($transaccion->motivo_id);
-                if (!$noGeneraComprobante) {
+                $no_genera_comprobante = TransaccionBodega::verificarMotivosEgreso($transaccion->motivo_id);
+                if (!$no_genera_comprobante) {
                     //creamos el comprobante
                     $transaccion->comprobante()->save(new Comprobante(['transaccion_id' => $transaccion->id]));
                     //lanzar el evento de la notificación
-                    $msg = 'Se ha generado un despacho de materiales a tu nombre, con transacción N°' . $transaccion->id . ', solicitado por ' . $modelo->solicitante->nombres . ' ' . $modelo->solicitante->apellidos . '. Por favor verifica y firma el movimiento';
+                    $msg = 'Se ha generado un despacho de materiales a tu nombre, con transacción N°' . $transaccion->id . ', solicitado por ' . Empleado::extraerNombresApellidos($transaccion->solicitante) . '. Por favor verifica y firma el movimiento';
                     event(new TransaccionEgresoEvent($msg, $url, $transaccion, false));
                 }
             }
@@ -210,20 +208,20 @@ class TransaccionBodegaEgresoController extends Controller
     public function update(TransaccionBodegaRequest $request, TransaccionBodega $transaccion)
     {
         $datos = $request->validated();
-        !is_null($request->pedido) ?? $datos['pedido_id'] = $request->safe()->only(['pedido'])['pedido'];
-        if ($request->transferencia) $datos['transferencia_id'] = $request->safe()->only(['transferencia'])['transferencia'];
-        if ($request->motivo) $datos['motivo_id'] = $request->safe()->only(['motivo'])['motivo'];
-        $datos['solicitante_id'] = $request->safe()->only(['solicitante'])['solicitante'];
-        $datos['sucursal_id'] = $request->safe()->only(['sucursal'])['sucursal'];
-        $datos['motivo_id'] = $request->safe()->only(['motivo'])['motivo'];
-        $datos['per_autoriza_id'] = $request->safe()->only(['per_autoriza'])['per_autoriza'];
-        if ($request->proyecto) $datos['proyecto_id'] = $request->safe()->only(['proyecto'])['proyecto'];
-        if ($request->etapa) $datos['etapa_id'] = $request->safe()->only(['etapa'])['etapa'];
-        if ($request->tarea) $datos['tarea_id'] = $request->safe()->only(['tarea'])['tarea'];
-        if ($request->per_atiende) $datos['per_atiende_id'] = $request->safe()->only(['per_atiende'])['per_atiende'];
+//        !is_null($request->pedido) ?? $datos['pedido_id'] = $request->safe()->only(['pedido'])['pedido'];
+//        if ($request->transferencia) $datos['transferencia_id'] = $request->safe()->only(['transferencia'])['transferencia'];
+//        if ($request->motivo) $datos['motivo_id'] = $request->safe()->only(['motivo'])['motivo'];
+//        $datos['solicitante_id'] = $request->safe()->only(['solicitante'])['solicitante'];
+//        $datos['sucursal_id'] = $request->safe()->only(['sucursal'])['sucursal'];
+//        $datos['motivo_id'] = $request->safe()->only(['motivo'])['motivo'];
+//        $datos['per_autoriza_id'] = $request->safe()->only(['per_autoriza'])['per_autoriza'];
+//        if ($request->proyecto) $datos['proyecto_id'] = $request->safe()->only(['proyecto'])['proyecto'];
+//        if ($request->etapa) $datos['etapa_id'] = $request->safe()->only(['etapa'])['etapa'];
+//        if ($request->tarea) $datos['tarea_id'] = $request->safe()->only(['tarea'])['tarea'];
+//        if ($request->per_atiende) $datos['per_atiende_id'] = $request->safe()->only(['per_atiende'])['per_atiende'];
 
-        $datos['autorizacion_id'] = $request->safe()->only(['autorizacion'])['autorizacion'];
-        $datos['estado_id'] = $request->safe()->only(['estado'])['estado'];
+//        $datos['autorizacion_id'] = $request->safe()->only(['autorizacion'])['autorizacion'];
+//        $datos['estado_id'] = $request->safe()->only(['estado'])['estado'];
 
         $transaccion->update($datos); //actualizar la transaccion
 
@@ -248,7 +246,6 @@ class TransaccionBodegaEgresoController extends Controller
 
             $modelo = new TransaccionBodegaResource($transaccion->refresh());
             $mensaje =  'Autorización actualizada correctamente';
-            return response()->json(compact('mensaje', 'modelo'));
         } else {
             if (auth()->user()->hasRole(User::ROL_BODEGA)) {
                 // Log::channel('testing')->info('Log', ['El bodeguero realiza la actualizacion?', true, $request->all(), 'datos: ', $datos]);
@@ -260,16 +257,17 @@ class TransaccionBodegaEgresoController extends Controller
                         $transaccion->estados()->attach($datos['estado_id']);
                     }
                     DB::commit();
-                } catch (Exception $e) {
+                } catch (Exception $ex) {
                     DB::rollBack();
-                    return response()->json(['mensaje' => 'Ha ocurrido un error al actualizar el registro'], 422);
+//                    return response()->json(['mensaje' => 'Ha ocurrido un error al actualizar el registro'], 422);
+                    throw Utils::obtenerMensajeErrorLanzable($ex);
                 }
             }
 
             $modelo = new TransaccionBodegaResource($transaccion->refresh());
             $mensaje = 'Estado actualizado correctamente';
-            return response()->json(compact('mensaje', 'modelo'));
         }
+        return response()->json(compact('mensaje', 'modelo'));
         // }
 
         /* $message = 'No tienes autorización para modificar esta solicitud';
@@ -294,19 +292,19 @@ class TransaccionBodegaEgresoController extends Controller
     {
         try {
             DB::beginTransaction();
-            $estadoAnulado = EstadoTransaccion::where('nombre', EstadoTransaccion::ANULADA)->first();
+            $estado_anulado = EstadoTransaccion::where('nombre', EstadoTransaccion::ANULADA)->first();
             $detalles = DetalleProductoTransaccion::where('transaccion_id', $transaccion->id)->get();
             foreach ($detalles as $detalle) {
-                $itemInventario = Inventario::find($detalle['inventario_id']);
-                $itemInventario->cantidad += $detalle['cantidad_inicial'];
-                $itemInventario->save();
-                $detalleProducto = DetalleProducto::find($itemInventario->detalle_id);
-                TransaccionBodega::activarDetalle($detalleProducto);
+                $item_inventario = Inventario::find($detalle['inventario_id']);
+                $item_inventario->cantidad += $detalle['cantidad_inicial'];
+                $item_inventario->save();
+                $detalle_producto = DetalleProducto::find($item_inventario->detalle_id);
+                TransaccionBodega::activarDetalle($detalle_producto);
                 if ($transaccion->pedido_id) {
-                    TransaccionBodega::restarDespachoPedido($transaccion->pedido_id, $itemInventario->detalle_id, $detalle['cantidad_inicial']);
+                    TransaccionBodega::restarDespachoPedido($transaccion->pedido_id, $item_inventario->detalle_id, $detalle['cantidad_inicial']);
                 }
             }
-            $transaccion->estado_id = $estadoAnulado->id;
+            $transaccion->estado_id = $estado_anulado->id;
             $transaccion->autorizacion_id = Autorizaciones::CANCELADO;
 
             $transaccion->save();
@@ -333,7 +331,7 @@ class TransaccionBodegaEgresoController extends Controller
         $modelo = $modelo->resolve();
         $modelo['listadoProductosTransaccion'] = $detalles;
 
-        return response()->json(compact('modelo'), 200);
+        return response()->json(compact('modelo'));
     }
 
     /**
@@ -348,18 +346,18 @@ class TransaccionBodegaEgresoController extends Controller
         $persona_retira = Empleado::find($transaccion->responsable_id);
         try {
             $transaccion = $resource->resolve();
-            $transaccion['listadoProductosTransaccion'] = TransaccionBodega::listadoProductos($transaccion['id']);;
+            $transaccion['listadoProductosTransaccion'] = TransaccionBodega::listadoProductos($transaccion['id']);
             // Log::channel('testing')->info('Log', ['Elementos a imprimir', ['transaccion' => $resource->resolve(), 'per_retira' => $persona_retira->toArray(), 'per_entrega' => $persona_entrega->toArray(), 'cliente' => $cliente]]);
             $pdf = Pdf::loadView('egresos.egreso', compact(['transaccion', 'persona_entrega', 'persona_retira', 'cliente', 'configuracion']));
             $pdf->setPaper('A5', 'landscape');
             $pdf->render();
-            $file = $pdf->output();
-            $filename = 'egreso_' . $resource->id . '_' . time() . '.pdf';
-            $ruta = storage_path() . DIRECTORY_SEPARATOR . 'app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . 'egresos' . DIRECTORY_SEPARATOR . $filename;
+            return $pdf->output();
+//            $filename = 'egreso_' . $resource->id . '_' . time() . '.pdf';
+//            $ruta = storage_path() . DIRECTORY_SEPARATOR . 'app' . DIRECTORY_SEPARATOR . 'public' . DIRECTORY_SEPARATOR . 'egresos' . DIRECTORY_SEPARATOR . $filename;
             // file_put_contents($ruta, $file); //en caso de que se quiera guardar el documento en el backend
-            return $file;
         } catch (Exception $ex) {
             Log::channel('testing')->info('Log', ['ERROR', $ex->getMessage(), $ex->getLine()]);
+            throw Utils::obtenerMensajeErrorLanzable($ex);
         }
     }
 
@@ -370,14 +368,12 @@ class TransaccionBodegaEgresoController extends Controller
     {
         $configuracion = ConfiguracionGeneral::first();
         $results = [];
-        $registros = [];
         switch ($request->accion) {
             case 'excel':
                 $results = $this->servicio->filtrarEgresoPorTipoFiltro($request);
                 $registros = TransaccionBodega::obtenerDatosReporteEgresos($results);
 
                 return Excel::download(new TransaccionBodegaEgresoExport(collect($registros)), 'reporte.xlsx');
-                break;
             case 'pdf':
                 try {
                     $results = $this->servicio->filtrarEgresoPorTipoFiltro($request);
@@ -424,7 +420,7 @@ class TransaccionBodegaEgresoController extends Controller
     /**
      * Esta función filtra las transacciones segun el estado de su comprobante y las envía al front donde se ubican en sus respectivas pestañas
      */
-    public function filtrarComprobante(Request $request)
+    public function filtrarComprobante()
     {
         $datos = TransaccionBodega::with('comprobante')->where('responsable_id', auth()->user()->empleado->id)
             ->whereHas('comprobante', function ($q) {
@@ -435,8 +431,9 @@ class TransaccionBodegaEgresoController extends Controller
         return response()->json(compact('results'));
     }
 
-    public function filtrarEgresos(Request $request)
+    public function filtrarEgresos()
     {
+        $datos = [];
         if (auth()->user()->hasRole([User::ROL_BODEGA, User::ROL_CONTABILIDAD, User::ROL_COORDINADOR, User::ROL_GERENTE, User::ROL_JEFE_TECNICO, User::ROL_EMPLEADO])) {
             $datos = TransaccionBodega::whereHas('comprobante', function ($q) {
                 $q->where('estado', request('estado'));
@@ -458,10 +455,10 @@ class TransaccionBodegaEgresoController extends Controller
                     $this->servicio->modificarItemEgresoParcial($request);
                     break;
             }
-            
+
 
             $modelo = [];
-        } catch (\Throwable $th) {
+        } catch (Throwable $th) {
             throw Utils::obtenerMensajeErrorLanzable($th);
         }
         return response()->json(compact('modelo'));
