@@ -26,50 +26,58 @@ class PermisoEmpleadoRequest extends FormRequest
     {
         return [
             'tipo_permiso' => 'required',
-            'fecha_hora_inicio' => 'required|string',
-            'fecha_hora_fin' => 'required|string',
+            'fecha_hora_inicio' => 'required|date',
+            'fecha_hora_fin' => 'required|date',
             'fecha_recuperacion' => 'nullable|date_format:Y-m-d',
             'hora_recuperacion' => 'nullable|string',
             'justificacion' => 'required|string',
-            'observacion'=> 'nullable|string',
-            'fecha_hora_reagendamiento'=> 'nullable|string',
+            'observacion' => 'nullable|string',
+            'fecha_hora_reagendamiento' => 'nullable|string',
             'empleado_id' => 'nullable|exists:empleados,id',
             'estado' => 'nullable',
             'tieneDocumento' => 'required',
             'cargo_vacaciones' => 'nullable',
-            'aceptar_sugerencia' => 'nullable'
+            'aceptar_sugerencia' => 'nullable',
+            'recupero' => 'boolean'
         ];
     }
+
     protected function prepareForValidation()
     {
-        $fecha_inicio = Carbon::createFromFormat('d-m-Y H:i', $this->fecha_hora_inicio);
-        $fecha_fin = Carbon::createFromFormat('d-m-Y H:i', $this->fecha_hora_fin);
-       if ($this->fecha_hora_reagendamiento != null) {
-        $fecha_hora_reagendamiento = Carbon::createFromFormat('d-m-Y H:i',$this->fecha_hora_reagendamiento);
-       }
-        if($this->fecha_recuperacion !=null) {
-        $fecha_recuperacion = $this->fecha_recuperacion;
+//        $mask = 'Y-m-d H:i:s';
+        $controller_method = $this->route()->getActionMethod();
+        $fecha_inicio = Carbon::parse($this->fecha_hora_inicio);
+        $fecha_fin = Carbon::parse($this->fecha_hora_fin);
+        if ($this->fecha_hora_reagendamiento != null) {
+//            $fecha_hora_reagendamiento = Carbon::createFromFormat('d-m-Y H:i', $this->fecha_hora_reagendamiento);
+//            $fecha_hora_reagendamiento = Carbon::createFromFormat($mask, $this->fecha_hora_reagendamiento);
+            $fecha_hora_reagendamiento = Carbon::parse($this->fecha_hora_reagendamiento);
+        }
+        if ($this->fecha_recuperacion != null) {
+            $fecha_recuperacion = $this->fecha_recuperacion;
         }
         //usuario logueado
-        if($this->empleado == null){
+        if ($this->empleado == null) {
             $this->merge([
                 'empleado_id' => auth()->user()->empleado->id,
             ]);
         }
-
-        $this->merge([
-            'fecha_hora_inicio' =>  $fecha_inicio->format('Y-m-d H:i:s'),
-            'fecha_hora_fin' =>  $fecha_fin->format('Y-m-d H:i:s'),
-        ]);
-        if($this->fecha_hora_reagendamiento != null )
-        {
+        if ($controller_method == 'store')
             $this->merge([
-             'fecha_hora_reagendamiento' => $fecha_hora_reagendamiento->format('Y-m-d H:i:s'),
+                'fecha_hora_inicio' => $fecha_inicio,
+                'fecha_hora_fin' => $fecha_fin,
+            ]);
+
+//        Log::channel('testing')->info('Log', ['PrepareForValidation', gettype($this->fecha_hora_inicio), gettype($this->fecha_hora_fin)]);
+
+        if ($this->fecha_hora_reagendamiento != null) {
+            $this->merge([
+                'fecha_hora_reagendamiento' => $fecha_hora_reagendamiento ?? null,
             ]);
         }
-        if($this->fecha_recuperacion !=null) {
+        if ($this->fecha_recuperacion != null) {
             $this->merge([
-                'fecha_recuperacion' => $fecha_recuperacion,
+                'fecha_recuperacion' => $fecha_recuperacion ?? null,
             ]);
         }
 
