@@ -4,7 +4,7 @@ namespace App\Http\Requests\RecursosHumanos\NominaPrestamos;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class PrestamoQuirorafarioRequest extends FormRequest
+class PrestamoQuirografarioRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
