@@ -2,25 +2,26 @@
 
 namespace App\Http\Resources\RecursosHumanos\NominaPrestamos;
 
+use App\Models\Empleado;
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class PrestamoQuirorafarioResource extends JsonResource
+class PrestamoQuirografarioResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
      */
     public function toArray($request)
     {
-        $modelo = [
+        return [
             'id' => $this->id,
             'mes' => $this->mes,
             'nut' => $this->nut,
-            'valor' =>  $this->valor,
+            'valor' => $this->valor,
             'empleado' => $this->empleado_id,
-            'empleado_info' => $this->empleado_info?$this->empleado_info->nombres.' '.$this->empleado_info->apellidos:'',        ];
-        return $modelo;
+            'empleado_info' => Empleado::extraerNombresApellidos($this->empleado),
+        ];
     }
 }
