@@ -1,35 +1,33 @@
 <?php
 
-use App\Http\Controllers\EmpleadoController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\ConceptoIngresoController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\DescuentosGeneralesController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\DescuentosLeyController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\ExtensionCoverturaSaludController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\MultaController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PrestamoHipotecarioController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PrestamoQuirirafarioController;
-use App\Http\Controllers\RecursosHumanos\NominaPrestamos\RolPagosController;
 use App\Http\Controllers\DepartamentoController;
+use App\Http\Controllers\EmpleadoController;
 use App\Http\Controllers\FondosRotativos\Saldo\SaldoGrupoController;
-use App\Http\Controllers\LoginSocialNetworkController;
 use App\Http\Controllers\RecursosHumanos\Alimentacion\AlimentacionController;
 use App\Http\Controllers\RecursosHumanos\Alimentacion\AsignarAlimentacionController;
 use App\Http\Controllers\RecursosHumanos\Alimentacion\DetalleAlimentacionController;
 use App\Http\Controllers\RecursosHumanos\AreasController;
 use App\Http\Controllers\RecursosHumanos\BancoController;
-use App\Http\Controllers\RecursosHumanos\EstadoCivilController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\ConceptoIngresoController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\DescuentosGeneralesController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\DescuentosLeyController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\EgresoRolPagoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\EstadoPermisoEmpleadoController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\ExtensionCoverturaSaludController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\FamiliaresControler;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\HorasExtrasSubTipoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\HorasExtrasTipoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\IngresoRolPagoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\LicenciaEmpleadoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\MotivoPermisoEmpleadoController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\MultaController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PeriodoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PermisoEmpleadoController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PrestamoEmpresarialController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PrestamoHipotecarioController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\PrestamoQuirografarioController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\RolPagoMesController;
+use App\Http\Controllers\RecursosHumanos\NominaPrestamos\RolPagosController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\SolicitudPrestamoEmpresarialController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\TipoLicenciaController;
 use App\Http\Controllers\RecursosHumanos\NominaPrestamos\VacacionController;
@@ -52,8 +50,8 @@ Route::apiResources(
         'descuentos_ley' => DescuentosLeyController::class,
         'multa' => MultaController::class,
         'prestamo_hipotecario' => PrestamoHipotecarioController::class,
-        'prestamo_quirorafario' => PrestamoQuirirafarioController::class,
-        'extension_covertura_salud' => ExtensionCoverturaSaludController::class,
+        'prestamos_quirografarios' => PrestamoQuirografarioController::class,
+        'extension_cobertura_salud' => ExtensionCoverturaSaludController::class,
         'prestamo_empresarial' => PrestamoEmpresarialController::class,
         'rubros' => RubroController::class,
         'solicitud_prestamo_empresarial' => SolicitudPrestamoEmpresarialController::class,
@@ -61,7 +59,6 @@ Route::apiResources(
         'tipo_licencia' => TipoLicenciaController::class,
         'licencia_empleado' => LicenciaEmpleadoController::class,
         'vacacion' => VacacionController::class,
-        // 'estado_civil' => EstadoCivilController::class,
         'areas' => AreasController::class,
         'familiares' => FamiliaresControler::class,
         'banco' => BancoController::class,
@@ -81,8 +78,13 @@ Route::apiResources(
     [
         'parameters' => [
             'descuentos_generales' => 'descuento_general',
+            'extension_cobertura_salud' => 'extension',
             'descuentos_ley' => 'descuento_ley',
-            'tipos_puestos_trabajos' => 'tipo_puesto_trabajo'
+            'prestamo_empresarial' => 'prestamo',
+            'prestamos_quirografarios' => 'prestamo',
+            'permiso_empleado' => 'permiso',
+            'tipos_puestos_trabajos' => 'tipo_puesto_trabajo',
+            'rol_pago_mes'=>'rol'
         ],
 
     ]
@@ -96,38 +98,38 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('prestamos_hipotecario_empleado', [PrestamoHipotecarioController::class, 'prestamos_hipotecario_empleado']);
     Route::put('aprobar_prestamo_empresarial', [SolicitudPrestamoEmpresarialController::class, 'aprobar_prestamo_empresarial']);
     Route::put('rechazar_prestamo_empresarial', [SolicitudPrestamoEmpresarialController::class, 'rechazar_prestamo_empresarial']);
-    Route::get('prestamos_quirorafario_empleado', [PrestamoQuirirafarioController::class, 'prestamos_quirorafario_empleado']);
-    Route::get('extension_covertura_salud_empleado', [ExtensionCoverturaSaludController::class, 'extension_covertura_salud_empleado']);
+    Route::get('$prestamos_quirografario_empleado', [PrestamoQuirografarioController::class, 'prestamoQuirografarioEmpleado']);
+    Route::get('extension_cobertura_salud_empleado', [ExtensionCoverturaSaludController::class, 'extensionCoberturaSaludEmpleado']);
     Route::get('sueldo_basico', [RubroController::class, 'sueldo_basico']);
     Route::get('porcentaje_iess', [RubroController::class, 'porcentaje_iess']);
     Route::get('porcentaje_anticipo', [RubroController::class, 'porcentaje_anticipo']);
-    Route::post('archivo_permiso_empleado', [PermisoEmpleadoController::class, 'archivo_permiso_empleado']);
+    Route::post('archivo_permiso_empleado', [PermisoEmpleadoController::class, 'archivoPermisoEmpleado']);
     Route::post('archivo_licencia_empleado', [LicenciaEmpleadoController::class, 'archivo_licencia_empleado']);
-    Route::get('archivo_permiso_empleado', [PermisoEmpleadoController::class, 'index_archivo_permiso_empleado']);
+    Route::get('archivo_permiso_empleado', [PermisoEmpleadoController::class, 'indexArchivoPermisoEmpleado']);
     Route::get('archivo_licencia_empleado', [LicenciaEmpleadoController::class, 'index_archivo_licencia_empleado']);
     Route::get('archivo_rol_pago', [RolPagosController::class, 'index_archivo_rol_pago_empleado']);
     Route::post('archivo_prestamo_hipotecario', [PrestamoHipotecarioController::class, 'archivo_prestamo_hipotecario']);
-    Route::post('archivo_prestamo_quirorafario', [PrestamoQuirirafarioController::class, 'archivo_prestamo_quirorafario']);
+    Route::post('archivo_prestamo_quirorafario', [PrestamoQuirografarioController::class, 'archivoPrestamoQuirografario']);
     Route::post('archivo_rol_pago', [RolPagosController::class, 'archivo_rol_pago_empleado']);
-    Route::post('archivo_extencion_conyugal', [ExtensionCoverturaSaludController::class, 'archivo_extension_conyugal']);
+    Route::post('archivo_extencion_conyugal', [ExtensionCoverturaSaludController::class, 'archivoExtensionConyugal']);
     Route::post('archivo-rol-pago-mes', [RolPagoMesController::class, 'importarRolPago']);
     Route::get('nivel_endeudamiento', [RolPagosController::class, 'nivel_endeudamiento']);
     Route::get('descuentos_permiso', [VacacionController::class, 'descuentos_permiso']);
-    Route::get('permisos_sin_recuperar', [PermisoEmpleadoController::class, 'permisos_sin_recuperar']);
-    Route::get('obtener_prestamo_empleado', [PrestamoEmpresarialController::class, 'obtener_prestamo_empleado']);
+    Route::get('permisos_sin_recuperar', [PermisoEmpleadoController::class, 'permisosSinRecuperar']);
+    Route::get('obtener_prestamo_empleado', [PrestamoEmpresarialController::class, 'obtenerPrestamoEmpleado']);
     Route::get('otener_saldo_empleado_mes', [SaldoGrupoController::class, 'otener_saldo_empleado_mes']);
     Route::get('imprimir_rol_pago/{rolPagoId}', [RolPagosController::class, 'imprimir_rol_pago']);
-    Route::get('imprimir_rol_pago_general/{rolPagoId}', [RolPagoMesController::class, 'imprimir_rol_pago_general']);
+    Route::get('imprimir_rol_pago_general/{rolPagoId}', [RolPagoMesController::class, 'imprimirRolPagoGeneral']);
     Route::get('imprimir_reporte_general_empleado', [EmpleadoController::class, 'imprimir_reporte_general_empleado']);
     Route::get('verificar-todos_roles-finalizadas', [RolPagoMesController::class, 'verificarTodasRolesFinalizadas']);
-    Route::get('finalizar-rol-pago', [RolPagoMesController::class, 'FinalizarRolPago']);
+    Route::get('finalizar-rol-pago', [RolPagoMesController::class, 'finalizarRolPago']);
     Route::get('habilitar-empleado', [EmpleadoController::class, 'HabilitaEmpleado']);
-    Route::get('imprimir_reporte_general/{rolPagoId}', [RolPagoMesController::class, 'imprimir_reporte_general']);
+    Route::get('imprimir_reporte_general/{rolPagoId}', [RolPagoMesController::class, 'imprimirReporteGeneral']);
     Route::get('enviar-roles-pago/{rolPagoId}',[RolPagoMesController::class, 'enviarRoles']);
     Route::get('enviar-rol-pago-empleado/{rolPagoId}',[RolPagosController::class, 'enviar_rolPago_empleado']);
-    Route::get('crear-cash-roles-pago/{rolPagoId}',[RolPagoMesController::class, 'crear_cash_rol_pago']);
-    Route::get('actualizar-rol-pago/{rolPagoId}',[RolPagoMesController::class, 'refrescar_rol_pago']);
-    Route::get('agregar-nuevos-empleados/{rol}',[RolPagoMesController::class, 'agregar_nuevos_empleados']);
+    Route::get('crear-cash-roles-pago/{rolPagoId}',[RolPagoMesController::class, 'crearCashRolPago']);
+    Route::get('actualizar-rol-pago/{rol}',[RolPagoMesController::class, 'refrescarRolPago']);
+    Route::get('agregar-nuevos-empleados/{rol}',[RolPagoMesController::class, 'agregarNuevosEmpleados']);
     Route::get('generar-username',[EmpleadoController::class, 'obtenerNombreUsuario']);
     Route::post('anular-prestamo-empresarial',[PrestamoEmpresarialController::class, 'deshabilitarPrestamo']);
     Route::get('crear-cash-alimentacion/{alimentacion_id}',[AlimentacionController::class, 'crear_cash_alimentacion']);
